Writing and reading secondary index files
A DiskRowSet's CFileSet has readers and writers for the non-core
data structures like the bloom filter and the ad hoc index. The
same should be done for secondary indexes.

The TabletComponents struct contains a RowSetTree that points to
numerous DiskRowSets, indexing them for search based on primary key.
TabletComponents should also contain SecondaryIndexes to handle the
top-level RowSet pruning (currently only done as removing the proper
for key ranges).

During Tablet::CaptureConsistentIterators(), we consult the
RowSetTree (FindRowSetsIntersectingInterval()) and NewRowIterator is
called using returned sets, which is a MaterializingRowIterator.

There should instead be a SecondaryIndexIterator that materializes only
the necessary blocks. This would entail modifying the Scan process.
Simpler would be to modify how Evaluate() is called and make that use
the secondary index.