Multi-Tablet checkpoint
Testing using multiple tablets has proven ineffective, whether due
to having shared_ptr complaining, or to UPSERTs not inserting to
CFiles, or to the necessity of unique keys for a single FsManager
and Tablet.

Instead, move to an implementation that creates one tablet per test.
This will not be as easily tunable, but it should do the trick.