Move towards opening the dictionary only once
Each block shares a dictionary block, "owned" by the CFileIterator.
Each dict block is constructed with this dictionary block as an input,
meaning it is shared among all of the dict blocks. This means that they
all have the same set of codewords that should be included in a given
predicate. Iterators are created with a single predicate in mind, so
these codewords can be preprocessed before scanning.