Fixed sel-vec behavior to match normal evaluation
In normal evaluation, selection vector starts out as full and cells
are cleared as evaluations occur. Once the block is materialized,
the result collector goes through the selection vector and fetches
only the data marked as full. This means that cleared elements are
completely ignored and do not have to be copied over.