Patch for tablet tests, deltafiles, plotting
Tablet tests were changed to support creating multiple tablets per
test. This has been reverted, as it is no longer used. Additionally,
tablet-decoder-eval-test has been modified to work with nrepeats and
a flag to run with pushdown on/off.

Deltafile iterators need to be checked to determine whether or not
to use predicate pushdown. No DMS or delta files is indicative of
a dataset without updates.

Plotting has been modified to compare the old implementation with
the pushdown implementation.