Resolve issues with eval_complete
Previously had been using eval_complete to signify whether or not a
single block could evaluate the predicate (e.g. for kCodeWordMode,
the block could evaluate, whereas for kPlainBinaryMode, the block
could not). This behavior does not work when there are different
block types for a given encodings.

Since the eval_complete flag is shared among the entire column's
blocks, individual blocks should not disagree with each other. This
would potentially lead to the column trying to evaluate all of its
data, even though some of the data may have been unset, since it was
evaluated to be unnecessary.