Fixed bug that resulted in incorrect rows
Had been clearing bits from the selection vector early before checking
that the number of bits to be cleared was below the number in the row.
This led to issues where if the block's dictionary did not satisfy the
predicate, extra bits past the end of the given selection vector view
were cleared, resulting in an lower value.

The fix to this was to add a function to the decoders to skip forward
by a given number of rows and to clear only as many bits as allowed
by the given decoder block.